small change - level now represents empirically height for elevator
