+++ src/test/java/org/sciborgs1155/robot/ElevatorTest.java
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.sciborgs1155.robot.Constants.Field.Level;
 import org.sciborgs1155.robot.elevator.Elevator;
+import org.sciborgs1155.robot.elevator.ElevatorConstants.Level;
 import org.sciborgs1155.robot.elevator.SimElevator;
 
 public class ElevatorTest {
@@ -38,10 +38,10 @@ public class ElevatorTest {
   private static Stream<Arguments> providePositionValues() {
     return Stream.of(
         Arguments.of(MIN_HEIGHT),
-        Arguments.of(Level.L1.height),
-        Arguments.of(Level.L2.height),
-        Arguments.of(Level.L3.height),
-        Arguments.of(Level.L4.height),
+        Arguments.of(Level.L1.extension),
+        Arguments.of(Level.L2.extension),
+        Arguments.of(Level.L3.extension),
+        Arguments.of(Level.L4.extension),
         Arguments.of(MAX_HEIGHT));
   }
 }
